feat(crystallize-ml): move experiment selection to screen
